+++ CHANGELOG.md
@@ -11,8 +11,8 @@
 
 ### Features
 
+- Profile envelopes are sent directly from profiler ([#2298](https://github.com/getsentry/sentry-java/pull/2298))
 - Add support for using Encoder with logback.SentryAppender ([#2246](https://github.com/getsentry/sentry-java/pull/2246))
-- Add captureProfile method to hub and client ([#2290](https://github.com/getsentry/sentry-java/pull/2290))
 - Report Startup Crashes ([#2277](https://github.com/getsentry/sentry-java/pull/2277))
 
 ### Dependencies